Move all pathfinding-related code into new file
